Update dating this branch with syntax corrections.
git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/mizzi@12768 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: bcefb4e0d920ef4364848285c978135756fdebad

+++ observations/MOPITT_CO/mopitt_ascii_to_obs.f90
@@ -143,7 +143,7 @@ real                            :: bin_beg, bin_end
 real                            :: sec, lat, lon, nlevels
 real                            :: pi ,rad2deg, re, wt, corr_err, fac, fac_obs_error
 real                            :: ln_10, xg_sec_avg, co_log_max, co_log_min, co_min
-real                            :: irot, nlvls_fix
+integer                         :: irot, nlvls_fix
 real*8, dimension(1000)         :: unif
 real*8, dimension(num_qc)       :: co_qc
 real*8, dimension(mop_dim)      :: co_avgker
